Updated file structure
